Agrego clases en Dominio

+++ src/main/java/com/tallerwebi/infraestructura/RepositorioLibro.java
@@ -1,11 +1,11 @@
 package com.tallerwebi.infraestructura;
 
-import com.tallerwebi.dominio.Libro.Libro;
+import com.tallerwebi.dominio.libro.Libro;
 
 import java.util.Set;
 
 public interface RepositorioLibro {
     public Set<Libro> getLibros();
 
-    public Libro getLibro(Integer id);
+    public Libro getLibro(Long id);
 }
