+++ src/test/java/com/tallerwebi/dominio/ServicioLibroTest.java
@@ -1,22 +1,18 @@
 package com.tallerwebi.dominio;
 
 
-import com.tallerwebi.dominio.Libro.Libro;
-import com.tallerwebi.dominio.Libro.ServicioLibro;
-import com.tallerwebi.dominio.Libro.ServicioLibroImpl;
+import com.tallerwebi.dominio.libro.Libro;
+import com.tallerwebi.dominio.libro.ServicioLibro;
 import com.tallerwebi.infraestructura.RepositorioLibro;
-import com.tallerwebi.infraestructura.RepositorioLibroImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import static org.hamcrest.core.IsNot.not;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.IsInstanceOf.instanceOf;
 import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -24,7 +20,7 @@ import static org.mockito.Mockito.when;
 
 public class ServicioLibroTest {
 
-    public static final Integer ID = 3;
+    public static final Long ID = 3L;
     private ServicioLibro servicioLibro;
     private RepositorioLibro repositorioLibro;
 
@@ -38,11 +34,11 @@ public class ServicioLibroTest {
     public void queSePuedaObtenerUnaListaDeLibrosQueNoEsteVacia(){
 
         Set libros = new HashSet<>();
-        libros.add(new Libro(1));
-        libros.add(new Libro(2));
-        libros.add(new Libro(3));
-        libros.add(new Libro(4));
-        libros.add(new Libro(5));
+        libros.add(new Libro(1L));
+        libros.add(new Libro(2L));
+        libros.add(new Libro(3L));
+        libros.add(new Libro(4L));
+        libros.add(new Libro(5L));
 
         when(this.servicioLibro.getLibros()).thenReturn(libros);
 
@@ -55,11 +51,11 @@ public class ServicioLibroTest {
     public void queSePuedaObtenerUnaListaDeLibrosQueDevuelvaCincoLibros(){
 
         Set libros = new HashSet<>();
-        libros.add(new Libro(1));
-        libros.add(new Libro(2));
-        libros.add(new Libro(3));
-        libros.add(new Libro(4));
-        libros.add(new Libro(5));
+        libros.add(new Libro(1L));
+        libros.add(new Libro(2L));
+        libros.add(new Libro(3L));
+        libros.add(new Libro(4L));
+        libros.add(new Libro(5L));
 
         when(this.servicioLibro.getLibros()).thenReturn(libros);
 
@@ -77,4 +73,6 @@ public class ServicioLibroTest {
 
         assertThat(libroObtenido.getID(), is(ID));
     }
+
+
 }
